initial solution
