+++ src/components/TodoItem.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable jsx-a11y/control-has-associated-label */
 import React, { useState, useEffect, useRef } from 'react';
 import classNames from 'classnames';
-import { Todo } from './types/Todo';
+import { Todo } from '../types/Todo';
 
 type Props = {
   todo: Todo,
